pyproject.toml
[tool.poetry]
name = "short-links-project"
version = "0.1.0"
description = ""
authors = ["it_tr1p"]
readme = "README.md"
packages = [{include = "aiogram_3x_template"}]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = {url = "https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip"}
sqlalchemy = "^2.0.15"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
greenlet = "^2.0.2"
python-dotenv = "^1.0.0"
pyshorteners = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
